AJAX Posting
polr.js will now prevent the form from doing a regular http submission and instead does it in the background, then updates the page. polr.jsp was modified to accomodate AJAX and to hide some buttons that are JavaScript dependent. PostProcessor was also modified to accept "captcha" and "g-recpatcha-response" parameters as the same thing.