DeveloperOptions / Dev Mode
Added a DeveloperOptions class that has one universal variable that can change the URL's on every page to match test environments. Each Java servlet will grab this variable and pass it to the .jsp's. The .jsp will incorporate it in its links and has a <meta> tag to inform JavaScript about the domain in use. All AJAX obeys the custom domain. The production domain name is "http://xarql.com".