Full replacement of /chat with a websocket
Made some small changes in the .java files to allow Message to have a more useful record of who sent it. Moved the websocket logic from chat.jsp to chat.js. Included a headers parser in chat.js. Added a status class to show server stats in a  different color on /chat.