Revised Database Interaction Classes
Removed the requirement for a httpServletResponse in the database interactors to reduce complexity. Changed the visibility of DatabaseInteractor's variables to reduce the chance of misuse by children of DatabaseQuery or DatabaseUpdate. Implemented AccountCreator as a child of DatabaseUpdate as a proof of concept. makeRequest() in DatabaseQuery and DatabaseUpdate will no longer throw an error; it will only return false if it can't make the request.